NEX-320 - Service ResultService has been added
